% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut.q.R
\name{cut.q}
\alias{cut.q}
\title{Cut variable into quantile}
\usage{
\method{cut}{q}(x, num.q = 4, probs)
}
\arguments{
\item{x}{Vector of quantitative data to transform}

\item{num.q}{Numer of categories to cut the variable into. Ignored if
probs is provided.}

\item{probs}{The probabilites at  wchich to cut the data into quantiles.}
}
\description{
Transform a quantitative variable into a factor by cutting it into
intervals. The breaks in those intervals are chosen by quantiles, at
regular probabilies, or at 'probs', if provided.
}
\examples{
x <- rnorm(100)
categorical <- cut.q(x, num.q=2) # catgorical is now a factor with 2 levels:
                             #  "[min(x), median(x)]", "[median(x), max(x)]"
}
\references{

}
\keyword{ecdf}
\keyword{quantile,}
