% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inverse.ecdf.R
\name{inverse.ecdf.conf.int}
\alias{inverse.ecdf.conf.int}
\title{Quantile function with confidence intervals}
\usage{
inverse.ecdf.conf.int(x, alpha, beta)
}
\arguments{
\item{x}{Vectored of ordered numerical data to calculate quntiles of}

\item{alpha}{The probabilites at which to calculate the quantiles}
}
\description{
Calculate quantiles, together with binomial-based confidence intervals.
}
\keyword{cdf}
\keyword{ecdf,}
