% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inverse.ecdf.R
\name{inverse.ecdf}
\alias{inverse.ecdf}
\title{Inverse Cumilative Density Function}
\usage{
inverse.ecdf(x)
}
\arguments{
\item{x}{Vector of ordered numerical data to generate the inv.cdf from}
}
\description{
Generate the inverse cdf, just as as 'cdf' function generates the 
cumilative density function.
}
\examples{
x <- rnorm(100)
F <- ecdf(x)
F.inf <- inverse.ecdf(x)
all(x == F.inv(F(x)))# == TRUE
}
\references{

}
\keyword{cdf}
\keyword{ecdf,}
