% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.glm.residuals.R
\name{plot.residuals}
\alias{plot.residuals}
\title{Plot glm residuals}
\usage{
\method{plot}{residuals}(model, model.name, residual.type = "pearson",
  response.type = "link", reest = FALSE)
}
\arguments{
\item{model}{Model object of which to find the objects}

\item{model.name}{The name to use in the plot}

\item{residual.type}{The type of residuals to extract, and plot}

\item{response.type}{The scale at which to show the predictions}

\item{reest}{Weather or not should be 'predict_reest' be used}
}
\description{
Plot glm residuals together with a kernel smoothing.
}
\keyword{cdf}
\keyword{ecdf,}
